traffic-ai/requirements.txt
supabase==2.4.0
python-dotenv==1.0.1
pandas==2.2.2
streamlit==1.35.0
plotly==5.22.0
